Check for null pointer returned by getLastVectorImageAtFrame where relevant
There are places that only have a Q_CHECK_PTR because they
follow a call to ScribbleArea::handleDrawingOnEmptyFrame which
should guarantee a frame after invocation. However this is called only
at the start of a stroke typically, which means that if the frame is
deleted while the stroke is in progress (ex. with a shortcut), then
it will still crash without a check. Users who aren't trying to break
the program will likely never do this, but we might as well add the
checks if we know the issue is there.

+++ core_lib/src/tool/erasertool.cpp
@@ -302,11 +302,12 @@ void EraserTool::removeVectorPaint()
     }
     else if (layer->type() == Layer::VECTOR)
     {
-        VectorImage*vectorImage = ((LayerVector*)layer)->getLastVectorImageAtFrame(mEditor->currentFrame(), 0);
+        mScribbleArea->clearBitmapBuffer();
+        VectorImage* vectorImage = static_cast<LayerVector*>(layer)->getLastVectorImageAtFrame(mEditor->currentFrame(), 0);
+        if (vectorImage == nullptr) { return; } // Can happen if the first frame is deleted while drawing
         // Clear the area containing the last point
         //vectorImage->removeArea(lastPoint);
         // Clear the temporary pixel path
-        mScribbleArea->clearBitmapBuffer();
         vectorImage->deleteSelectedPoints();
 
         mScribbleArea->setModified(mEditor->layers()->currentLayerIndex(), mEditor->currentFrame());
